Add MviUserAttributes. Make HCA::UserAttributes inherit MviUserAttributes. Update MVI::UserAttributes spec to test both
